requirements.txt
pydexcom==0.2.3
Flask
python-dotenv==0.21.0
requests==2.28.1
rgbxy==0.5
